requirements.txt
sty
pyyaml
asyncssh
urwid
aiohttp